Reorg registers into files
