+++ system_zero/src/lib.rs
@@ -5,9 +5,9 @@
 #![allow(clippy::diverging_sub_expression)]
 
 mod arithmetic;
-mod column_layout;
 mod core_registers;
 mod memory;
 mod permutation_unit;
 mod public_input_layout;
+mod registers;
 pub mod system_zero;
