+++ system_zero/src/permutation_unit.rs
@@ -8,9 +8,9 @@ use starky::constraint_consumer::{ConstraintConsumer, RecursiveConstraintConsume
 use starky::vars::StarkEvaluationTargets;
 use starky::vars::StarkEvaluationVars;
 
-use crate::column_layout::permutation::*;
-use crate::column_layout::NUM_COLUMNS;
 use crate::public_input_layout::NUM_PUBLIC_INPUTS;
+use crate::registers::permutation::*;
+use crate::registers::NUM_COLUMNS;
 use crate::system_zero::SystemZero;
 
 fn constant_layer<F, FE, P, const D2: usize>(
@@ -269,10 +269,10 @@ mod tests {
     use starky::constraint_consumer::ConstraintConsumer;
     use starky::vars::StarkEvaluationVars;
 
-    use crate::column_layout::permutation::{col_input, col_output};
-    use crate::column_layout::NUM_COLUMNS;
     use crate::permutation_unit::SPONGE_WIDTH;
     use crate::public_input_layout::NUM_PUBLIC_INPUTS;
+    use crate::registers::permutation::{col_input, col_output};
+    use crate::registers::NUM_COLUMNS;
     use crate::system_zero::SystemZero;
 
     #[test]
