+++ system_zero/src/registers/logic.rs
@@ -1,0 +1,3 @@
+//! Logic unit.
+
+pub(super) const END: usize = super::START_LOGIC;
