+++ system_zero/src/registers/memory.rs
@@ -1,0 +1,3 @@
+//! Memory unit.
+
+pub(super) const END: usize = super::START_MEMORY;
